moved password state out to PasswordValidation component
